Cleanup and reorganize code
