--- src/constants/logger-types.js
@@ -1,6 +1,0 @@
-const LOGGER_TYPES = {
-  INFO: 'INFO',
-  ERROR: 'ERROR'
-};
-
-exports.LOGGER_TYPES = LOGGER_TYPES;
