--- src/constants/status-codes.js
@@ -1,7 +1,0 @@
-const STATUS_CODES = {
-  END: 'e',
-  FULL: 'f',
-  CACHE: 'c'
-};
-
-exports.STATUS_CODES = STATUS_CODES;
