--- src/utils/chunk-string.js
@@ -1,13 +1,0 @@
-function chunkString(str, len) {
-  const size = Math.ceil(str.length / len);
-  const r = Array(size);
-  let offset = 0;
-
-  for (let i = 0; i < size; i++) {
-    r[i] = str.substr(offset, len);
-    offset += len;
-  }
-  return r;
-}
-
-exports.chunkString = chunkString;
